Finish ex19

+++ src/main/java/com/xpanxion/java/springboot/da1/demo/repository/student5/MemberRepository5.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberRepository5 extends JpaRepository<Member5, Integer> {
+public interface MemberRepository5 extends JpaRepository<Member5, Long> {
 }
